Move scripts endpoints to Fleet Free
