+++ server/service/orbit.go
@@ -502,11 +502,24 @@ func getOrbitScriptEndpoint(ctx context.Context, request interface{}, svc fleet.
 }
 
 func (svc *Service) GetHostScript(ctx context.Context, execID string) (*fleet.HostScriptResult, error) {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
+	// this is not a user-authenticated endpoint
 	svc.authz.SkipAuthorization(ctx)
 
-	return nil, fleet.ErrMissingLicense
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		return nil, fleet.OrbitError{Message: "internal error: missing host from request context"}
+	}
+
+	// get the script's details
+	script, err := svc.ds.GetHostScriptExecutionResult(ctx, execID)
+	if err != nil {
+		return nil, err
+	}
+	// ensure it cannot get access to a different host's script
+	if script.HostID != host.ID {
+		return nil, ctxerr.Wrap(ctx, newNotFoundError(), "no script found for this host")
+	}
+	return script, nil
 }
 
 /////////////////////////////////////////////////////////////////////////////////
@@ -543,11 +556,46 @@ func postOrbitScriptResultEndpoint(ctx context.Context, request interface{}, svc
 }
 
 func (svc *Service) SaveHostScriptResult(ctx context.Context, result *fleet.HostScriptResultPayload) error {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
+	// this is not a user-authenticated endpoint
 	svc.authz.SkipAuthorization(ctx)
 
-	return fleet.ErrMissingLicense
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		return fleet.OrbitError{Message: "internal error: missing host from request context"}
+	}
+	if result == nil {
+		return ctxerr.Wrap(ctx, &fleet.BadRequestError{Message: "missing script result"}, "save host script result")
+	}
+
+	// always use the authenticated host's ID as host_id
+	result.HostID = host.ID
+	hsr, err := svc.ds.SetHostScriptExecutionResult(ctx, result)
+	if err != nil {
+		return ctxerr.Wrap(ctx, err, "save host script result")
+	}
+
+	if hsr != nil {
+		var user *fleet.User
+		if hsr.UserID != nil {
+			user, err = svc.ds.UserByID(ctx, *hsr.UserID)
+			if err != nil {
+				return ctxerr.Wrap(ctx, err, "get host script execution user")
+			}
+		}
+		if err := svc.ds.NewActivity(
+			ctx,
+			user,
+			fleet.ActivityTypeRanScript{
+				HostID:            host.ID,
+				HostDisplayName:   host.DisplayName(),
+				ScriptExecutionID: hsr.ExecutionID,
+				Async:             !hsr.SyncRequest,
+			},
+		); err != nil {
+			return ctxerr.Wrap(ctx, err, "create activity for script execution request")
+		}
+	}
+	return nil
 }
 
 /////////////////////////////////////////////////////////////////////////////////
